Move security configuration from properties to Java code
Spring security related configuration has been removed from the application.properties file and shifted into the SecurityConfig.Java file. Instead of using properties, the application now defines users and roles with an InMemoryUserDetailsManager. A password encoder is also defined, though it uses the deprecated NoOpPasswordEncoder for now as this is just for local development.